chore:获取的size进行单位转换

+++ android/src/main/java/im/shimo/react/keyboard/KeyboardViewManager.java
@@ -88,7 +88,7 @@ public class KeyboardViewManager extends ViewGroupManager<KeyboardView> {
         return height;
     }
 
-    static int getNavigationSize() {
+    static float getNavigationSize() {
         if (INSTANCE == null || INSTANCE.mKeyboardView == null) {
             return 0;
         } else {
